creating comments is now a modal
