+++ frontend/src/components/CreateComment/CreateComment.css
@@ -1,0 +1,8 @@
+.comment-form-container {
+  margin-top: 10px;
+  margin-bottom: 10px;
+}
+
+#body {
+  
+}
