+++ frontend/src/components/Feed/Post.js
@@ -7,36 +7,15 @@ import { likePost } from '../../store/like';
 import PostModal from '../PostModal';
 import CommentModal from '../PostModal/commentExpand';
 import CommentTwoModal from '../PostModal/commentExpandTwo';
+import CreateComment from '../CreateComment/CreateComment';
 import { getComments, createComment, deleteComment } from '../../store/comment'
-import * as yup from 'yup';
-import { useFormik } from 'formik';
+
 
 function Post({ post }) {
   const dispatch = useDispatch();
   const user = useSelector((state) => state.session.user)
   const numberOfComments = post?.Comments?.length;
   const hasComments = numberOfComments > 0;
-
-  const like = async () => {
-    await dispatch(likePost(post.id))
-    await dispatch(getPosts())
-  }
-
-  const formik = useFormik({
-    initialValues: {
-      body: "",
-      userId: user.id,
-      postId: post.id
-
-    },
-    validationSchema: yup.object({
-      body: yup.string().min(5).max(350).required('Comment must be be between 5 and 350 characters'),
-    }),
-    onSubmit: async (values) => {
-      dispatch(createComment(values))
-      formik.values.body = "";
-    },
-  });
 
   const isLiked = () => {
     const likes = post.Likes;
@@ -51,6 +30,11 @@ function Post({ post }) {
     return false;
   };
 
+  const like = async () => {
+    await dispatch(likePost(post.id))
+    await dispatch(getPosts())
+  }
+  
   const unlike = async () => {
     await dispatch(likePost(post.id))
     await dispatch(getPosts())
@@ -66,10 +50,6 @@ function Post({ post }) {
       return <div className="last-comment-wrapper">{comment.body}</div>;
     }
   };
-
-  // let lastComment = post?.Comments[post.Comments.length - 1]
-
-  // console.log(lastComment, 'here')
 
   return (
     <div className='post'>
@@ -133,24 +113,7 @@ function Post({ post }) {
       <div>
         <span>{lastComment()}</span>
       </div>
-      <form onSubmit={formik.handleSubmit}>
-        <div className='postForm'>
-          <div>
-            <textarea
-              id='body'
-              name='body'
-              type='text'
-              onChange={formik.handleChange}
-              value={formik.values.body}
-              onBlur={formik.handleBlur}
-            />
-            {formik.touched.body && formik.errors.body ? (
-              <div className="errorText">{formik.errors.body}</div>
-            ) : null}
-          </div>
-            <button type="submit">Done</button>
-        </div>
-      </form>
+      <CreateComment post={post}/>
     </div>
   )
 }
